tests/requirements.txt
-i https://pypi.org/simple
ansible-base==2.10.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
ansible-runner-http==1.0.0
ansible-runner==1.4.6
ansible==2.10.0
arrow==0.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
bcrypt==3.2.0; python_version >= '3.6'
binaryornot==0.4.4
cachetools==4.1.1; python_version ~= '3.5'
cerberus==1.3.2; python_version >= '2.7'
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click-completion==0.5.2
click-help-colors==0.8
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
colorama==0.4.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
coloredlogs==14.0
cookiecutter==1.7.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cryptography==3.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
docker==4.3.1
docutils==0.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dotmap==1.3.17
fasteners==0.15
google-auth==1.22.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
html-testrunner==1.2.1
humanfriendly==8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
inflection==0.5.1
jinja2-time==0.2.0
jinja2==2.11.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
kubernetes==11.0.0
lockfile==0.12.2
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
molecule==3.0.8
monotonic==1.5
numpy==1.19.2; python_version >= '3.6'
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
openshift==0.11.2
packaging==20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pandas==1.1.3
paramiko==2.7.2
pathspec==0.8.0
pexpect==4.8.0
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
poyo==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
psutil==5.7.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
ptyprocess==0.6.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pynacl==1.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-daemon==2.2.4
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-gilt==1.2.3; python_version >= '3.6'
python-slugify==4.0.1
python-string-utils==1.0.0; python_version >= '3.5'
pytz==2020.1
pyyaml==5.3.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
requests==2.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rsa==4.6; python_version >= '3.5'
ruamel.yaml.clib==0.2.2; python_version < '3.9' and platform_python_implementation == 'CPython'
ruamel.yaml==0.16.12
sh==1.13.1
shellingham==1.3.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tabulate==0.8.7
text-unidecode==1.3
tree-format==0.1.2
urllib3==1.25.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
websocket-client==0.57.0
yamllint==1.25.0
